+++ app/src/main/java/com/gurumlab/vocaroutine/data/source/remote/GptApiClient.kt
@@ -2,54 +2,11 @@ package com.gurumlab.vocaroutine.data.source.remote
 
 import com.gurumlab.vocaroutine.data.model.ChatRequest
 import com.gurumlab.vocaroutine.data.model.ChatResponse
-import com.squareup.moshi.Moshi
-import okhttp3.Interceptor
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.moshi.MoshiConverterFactory
 import retrofit2.http.Body
 import retrofit2.http.POST
-import java.util.concurrent.TimeUnit
 
 interface GptApiClient {
 
     @POST("v1/chat/completions")
     suspend fun getResponse(@Body requestBody: ChatRequest): ChatResponse
-
-    companion object {
-        private val moshi = Moshi.Builder()
-            .build()
-
-        private const val BASE_URL = "https://api.openai.com/"
-
-        fun create(apiKey: String): GptApiClient {
-            val logger = HttpLoggingInterceptor().apply {
-                level = HttpLoggingInterceptor.Level.BODY
-            }
-
-            val header = Interceptor { chain ->
-                val newRequest = chain.request().newBuilder()
-                    .addHeader("Content-Type", "application/json")
-                    .addHeader("Authorization", "Bearer $apiKey")
-                    .build()
-                chain.proceed(newRequest)
-            }
-
-            val client = OkHttpClient.Builder()
-                .connectTimeout(1, TimeUnit.MINUTES)
-                .readTimeout(30, TimeUnit.SECONDS)
-                .writeTimeout(15, TimeUnit.SECONDS)
-                .addInterceptor(logger)
-                .addInterceptor(header)
-                .build()
-
-            return Retrofit.Builder()
-                .baseUrl(BASE_URL)
-                .client(client)
-                .addConverterFactory(MoshiConverterFactory.create(moshi))
-                .build()
-                .create(GptApiClient::class.java)
-        }
-    }
 }
